refactor: 여러 페이지에서 활용되는 컴포넌트 handleReview utils 아래로 이동
