+++ src/component/mypage/MyReview.tsx
@@ -1,6 +1,6 @@
 import InquireBoxTitle from "../utils/InquireBoxTitle";
 import Reserve from "../../asset/img/list-check-solid.svg";
-import HandleReview from "../book/review/HandleReview";
+import HandleReview from "../utils/HandleReview";
 import axiosPromise from "../../util/axios";
 import Pagination from "../utils/Pagination";
 import { useGetMyReviewInfo } from "../../api/reviews/useGetMyReviewInfo";
